feat(array): add nil value consistency handling
In Lua, if we set an index of an array as NIL `arr[n] = nil` and lets say `arr` has the size of 3, it will result in a hole in the array. { [1] = v, [2] = nil, [3] = n }
the index still exists and not removed from the array. This causes inconsistency. This commit ensures that whenever an index has a value of NIL, the index will be removed.